Refactor API to use ES modules in more descriptive structure
