+++ src/backend/api_components/middleware.mjs
@@ -1,0 +1,1 @@
+// TODO: CREATE MIDDLEWARE TO VERIFY JWTs.
